sandbox: production OS-level profiles + per-OS backend split (#1647)
Promote the production sandbox path from "use a worktree + capability
ceiling" to a layered defense-in-depth with explicit profile selection
and a per-OS backend trait.

`CapabilityPolicy` now carries a `sandbox_profile: SandboxProfile`
field with four variants:

- `unrestricted` — bypass everything (escape hatch)
- `worktree` *(default)* — workspace path enforcement plus
  best-effort OS confinement; honors `HARN_HANDLER_SANDBOX={off,warn,enforce}`
- `os_hardened` — required OS confinement; spawns return
  `tool_rejected` if the platform mechanism is missing, regardless of
  `HARN_HANDLER_SANDBOX`
- `wasi` — testbench-only

`CapabilityPolicy::intersect` always picks the strictest of the two
profiles so a lenient parent ceiling cannot weaken a child request.

The `host_call("process", "exec", ...)` host call accepts
`sandbox_profile: "..."` to promote a single spawn without rewriting
the surrounding policy.

Three new Harn builtins surface backend identity for diagnostics and
conformance fixtures: `sandbox_active_backend()`,
`sandbox_backend_available()`, and `sandbox_active_profile()`.

The 1.2k-line `crates/harn-vm/src/stdlib/sandbox.rs` was split into
a `sandbox/` directory with one impl per OS behind a shared
`SandboxBackend` trait — Linux Landlock+seccomp, macOS sandbox-exec,
Windows AppContainer+JobObject, and OpenBSD pledge+unveil. Adding a
backend is now one new file plus two `mod` lines in `mod.rs`; callers
no longer branch on `cfg(target_os)`.

Coverage:

- 8 new Rust integration tests in `crates/harn-vm/tests/sandbox_hardened.rs`
- 2 new conformance fixtures under
  `conformance/tests/sandbox_hardened/` (backend identity + profile
  override)
- Existing per-platform tests preserved

Documentation:

- New `docs/src/sandboxing.md` with the per-platform capability →
  kernel-knob mapping table, profile selection examples, and replay
  semantics
- README, `docs/src/host-boundary.md`,
  `docs/src/dev/platform-compatibility.md`, `spec/HARN_SPEC.md`, and
  `docs/src/language-spec.md` updated to match the new behavior
