+++ conformance/tests/sandbox_hardened/sandbox_hardened_profile_override.expected
@@ -1,0 +1,4 @@
+true
+ok
+true
+true
