+++ docs/theme/harn-keywords.js
@@ -9,5 +9,5 @@
 window.__HARN_KEYWORDS = {
   keyword: "ask_user break catch continue deadline defer dual_control else emit enum escalate_to eval_pack exclusive extends finally fn for from guard if impl import in interface let match mutex override parallel pipeline pub request_approval require retry return select skill spawn struct throw to tool try type var while yield",
   literal: "true false nil",
-  built_in: "Err Ok abs acos add_assistant add_message add_system add_tool_result add_user addr_of advance_time agent agent_budget agent_chat_loop agent_chat_route_input agent_chat_wait_for_user_tools agent_config agent_daemon_snapshot agent_daemon_step agent_dispatch_tool_batch agent_dispatch_tool_call agent_inject_feedback agent_llm_turn agent_loop agent_name agent_parse_tool_calls agent_preset agent_session_ancestry agent_session_claim_tool_format agent_session_close agent_session_compact agent_session_current_id agent_session_exists agent_session_fork agent_session_fork_at agent_session_inject agent_session_length agent_session_open agent_session_reset agent_session_seed_from_jsonl agent_session_snapshot agent_session_system_prompt agent_session_tool_format agent_session_trim agent_subscribe agent_trace agent_trace_summary agent_turn agent_typed_output_checkpoint agentic_user append_file arch artifact artifact_apply_intent artifact_command_result artifact_context artifact_derive artifact_diff artifact_diff_review artifact_editor_selection artifact_git_diff artifact_patch_proposal artifact_review_decision artifact_select artifact_test_result artifact_verification_bundle artifact_verification_result artifact_workspace_file artifact_workspace_snapshot asin assemble_context assert assert_eq assert_ne asset_root atan atan2 atomic atomic_add atomic_cas atomic_get atomic_set audit_agent await base32_decode base32_encode base64_decode base64_encode base64url_decode base64url_encode basename blake3 bold brotli_decode brotli_encode bytes_concat bytes_eq bytes_from_base64 bytes_from_hex bytes_from_string bytes_len bytes_slice bytes_to_base64 bytes_to_hex bytes_to_string bytes_to_string_lossy camel_to_kebab camel_to_pascal camel_to_snake cancel cancel_graceful capture_stderr_start capture_stderr_take ceil channel channel_select checkpoint checkpoint_clear checkpoint_delete checkpoint_exists checkpoint_get checkpoint_list chunk circuit_breaker circuit_check circuit_record_failure circuit_record_success circuit_reset clear_persona_hooks clear_tool_hooks close_agent close_channel color command_llm_risk_scan command_policy command_policy_pop command_policy_push command_result_scan command_risk_scan composition_binding_manifest composition_crystallization_trace composition_execute composition_search_examples composition_typescript_declarations compute_content_hash connector_call connector_shared_verify_jwt_inline constant_time_eq contains context_pack_manifest context_pack_manifest_parse context_pack_suggestions continue_as_new conversation cookie_delete cookie_parse cookie_round_trip cookie_serialize cookie_sign cookie_verify copy_file correction_query correction_record corrections.query corrections.record cos csv_parse csv_stringify cwd daemon_resume daemon_snapshot daemon_spawn daemon_stop daemon_trigger date_add date_diff date_format date_from_components date_in_zone date_iso date_now date_now_iso date_parse date_to_zone dedup_by delete_file dim dirname duration_days duration_hours duration_minutes duration_ms duration_seconds duration_to_human duration_to_seconds ed25519_keypair ed25519_sign ed25519_verify egress_policy elapsed enable_tracing ends_with entries env env_or eprint eprintln error_category estimate_tokens eval_metric eval_metrics eval_pack_manifest eval_pack_run eval_suite_manifest eval_suite_run event_log.emit event_log.latest event_log.subscribe event_log_emit exec exec_at execution_root exit exp extname file_exists fixture_user flat_map floor flow_evidence_atom flow_evidence_citation flow_evidence_metadata flow_evidence_transcript flow_invariant_allow flow_invariant_block flow_invariant_confidence flow_invariant_is_blocking flow_invariant_kind flow_invariant_require_approval flow_invariant_warn flow_remediation flow_with_confidence flow_with_evidence flow_with_remediation format friction_clear friction_eval_fixture friction_event friction_events friction_record git.conflicts git.diff git.fetch git.merge_base git.push git.rebase git.repo.discover git.status git.worktree.create git.worktree.remove glob group_by gzip_decode gzip_encode handler_context handoff handoff_context handoff_routes harn.mcp.roots hash_value hex_decode hex_encode hitl_pending hmac_sha1 hmac_sha256 hmac_sha256_base64 home_dir host_call host_capabilities host_has host_mock host_mock_calls host_mock_clear host_mock_pop_scope host_mock_push_scope host_tool_call host_tool_list hostname http_delete http_download http_get http_header http_mock http_mock_calls http_mock_clear http_patch http_post http_put http_request http_response http_response_bytes http_response_json http_response_text http_server http_server_after http_server_before http_server_on_shutdown http_server_readiness http_server_ready http_server_request http_server_route http_server_security_headers http_server_set_ready http_server_shutdown http_server_test http_server_tls_edge http_server_tls_pem http_server_tls_plain http_server_tls_self_signed_dev http_session http_session_close http_session_request http_stream_close http_stream_info http_stream_open http_stream_read invalidate_facts is_cancelled is_err is_infinite is_nan is_ok is_rate_limited is_same is_stderr_tty is_stdin_tty is_stdout_tty is_timeout is_type iter join jq jq_first json_extract json_parse json_pointer json_pointer_delete json_pointer_set json_stringify json_validate jwt_sign jwt_verify kebab_to_camel kebab_to_snake keys len list_agents list_dir list_providers_native llm_available_providers llm_budget llm_budget_remaining llm_call llm_call_safe llm_call_structured llm_call_structured_result llm_call_structured_safe llm_compare_costs llm_completion llm_config llm_cost llm_format_usd llm_healthcheck llm_infer_provider llm_info llm_known_models llm_mock llm_mock_calls llm_mock_clear llm_mock_pop_scope llm_mock_push_scope llm_model_defaults llm_model_info llm_model_tier llm_pick_model llm_pricing llm_provider_catalog llm_providers llm_qc_default_model llm_rate_limit llm_resolve_model llm_resolved_options llm_session_cost llm_stream llm_stream_call llm_usage ln load_run_tree load_skill log log10 log2 log_debug log_error log_info log_json log_set_level log_warn lowercase lowercase_first mailbox_close mailbox_lookup mailbox_metrics mailbox_open mailbox_receive mailbox_send mailbox_try_receive max mcp_call mcp_connect mcp_disconnect mcp_elicit mcp_ensure_active mcp_get_prompt mcp_list_prompts mcp_list_resource_templates mcp_list_resources mcp_list_tools mcp_prompt mcp_read_resource mcp_registry_status mcp_release mcp_report_progress mcp_resource mcp_resource_template mcp_roots mcp_serve mcp_server_card mcp_server_info mcp_tools md5 mean median merge_captain_agent metadata_entries metadata_get metadata_refresh_hashes metadata_resolve metadata_save metadata_set metadata_stale metadata_status metrics_inc microcompact min mkdir mock_stdin mock_time mock_tty monitor_wait_for_native monotonic_ms month_name move_file multipart_field_bytes multipart_field_text multipart_form_data multipart_parse now_ms oncall_captain_agent pair parallel_race parse_junit_xml partition pascal_to_camel pascal_to_snake path_basename path_extension path_is_absolute path_is_relative path_join path_normalize path_parent path_parts path_relative_to path_segments path_stem path_to_native path_to_posix path_with_extension path_with_stem path_workspace_info path_workspace_normalize percentile persona_eval_ladder_manifest persona_eval_ladder_run pg_close pg_connect pg_execute pg_mock_calls pg_mock_pool pg_pool pg_query pg_query_one pg_transaction pid plan_artifact plan_entries platform pow print println progress project_catalog_native project_enrich_native project_fingerprint project_root project_scan_native project_scan_tree_native project_walk_tree_native prompt_mark_rendered prompt_user provider_capabilities provider_capabilities_clear provider_capabilities_install provider_register query_parse query_stringify random random_choice random_int random_shuffle range read_file read_file_bytes read_file_result read_line read_lines read_stdin receive regex_captures regex_match regex_replace regex_replace_all regex_split register_persona_hook register_step_hook register_tool_hook release_captain_agent render render_always_on_catalog render_prompt render_string render_with_provenance repair_agent replace resume_agent review_captain_agent rng_seed round routing_policy run_record run_record_diff run_record_eval run_record_eval_suite run_record_fixture run_record_load run_record_save runtime_context runtime_context_clear runtime_context_get runtime_context_set runtime_context_values runtime_paths scan_directory schema_check schema_expect schema_extend schema_from_json_schema schema_from_openapi_schema schema_is schema_of schema_omit schema_parse schema_partial schema_pick schema_recover schema_to_json_schema schema_to_openapi_schema scripted_user secret_get secret_scan select_artifacts_adaptive self_review send send_input session_cookie session_from_cookies session_sign session_verify set set_add set_color_mode set_contains set_difference set_intersect set_is_disjoint set_is_subset set_is_superset set_remove set_symmetric_difference set_union sha224 sha256 sha384 sha3_256 sha3_512 sha512 sha512_256 shared_cas shared_cell shared_get shared_map shared_map_cas shared_map_delete shared_map_entries shared_map_get shared_map_set shared_map_snapshot shared_metrics shared_scope_id shared_set shared_snapshot shell shell_at sign signed_url simulated_user_post_turn simulated_user_read_tools simulated_user_respond simulated_user_status sin skill_count skill_define skill_describe skill_find skill_list skill_registry skill_remove skill_render skill_select skill_who_signed skills_catalog_entries sleep sleep_ms snake_to_camel snake_to_kebab snake_to_pascal source_dir spawn_agent split sqrt sse_close sse_connect sse_event sse_mock sse_receive sse_server_cancel sse_server_cancelled sse_server_close sse_server_disconnected sse_server_flush sse_server_heartbeat sse_server_mock_disconnect sse_server_mock_receive sse_server_response sse_server_send sse_server_status starts_with stat stddev store_clear store_delete store_get store_list store_save store_set str_pad stream.broadcast stream.collect stream.debounce stream.filter stream.first stream.fold stream.interleave stream.map stream.merge stream.race stream.scan stream.take stream.take_until stream.tap stream.throttle stream.zip sub_agent_request sub_agent_run substring summary_agent supervisor_events supervisor_metrics supervisor_start supervisor_state supervisor_stop sync_gate_acquire sync_metrics sync_mutex_acquire sync_release sync_rwlock_acquire sync_semaphore_acquire tan tar_create tar_extract task_current temp_dir testbench_clock_leaks testbench_fs_diff testbench_is_active throw_error tiktoken_count_tokens tiktoken_tokenizer_info timer_end timer_start timestamp title_case to_float to_int to_list to_string toml_parse toml_stringify tool_bind tool_count tool_def tool_define tool_describe tool_find tool_format_result tool_list tool_parse_call tool_prompt tool_ref tool_registry tool_remove tool_schema tool_select tool_surface_validate tool_synth_invoke tool_synthesis_cache tool_synthesis_clear tool_synthesize trace_end trace_id trace_spans trace_start trace_summary transcript transcript_abandon transcript_add_asset transcript_archive transcript_assets transcript_auto_compact transcript_compact transcript_events transcript_events_by_kind transcript_export transcript_fork transcript_from_messages transcript_id transcript_import transcript_messages transcript_render_full transcript_render_visible transcript_reset transcript_resume transcript_stats transcript_summarize transcript_summary transport_mock_calls transport_mock_clear trigger_fire trigger_inspect_action_graph trigger_inspect_dlq trigger_inspect_lifecycle trigger_list trigger_register trigger_replay trigger_test_harness trim trust.policy_for trust.query trust.record trust.score trust.verify_chain trust_graph_policy_for trust_graph_query trust_graph_record trust_graph_verify_chain trust_query trust_record try_receive type_of unicode_graphemes unicode_normalize unmock_stdin unmock_time unmock_tty unreachable unwrap unwrap_err unwrap_or uppercase uppercase_first url_build url_decode url_encode url_parse user_tools username uuid uuid_nil uuid_parse uuid_v5 uuid_v7 values variance verify_agent verify_signed_url vision_ocr wait_agent waitpoint_cancel waitpoint_complete waitpoint_create waitpoint_wait walk_dir webhook_intake_deregister webhook_intake_feed webhook_intake_list webhook_intake_recent webhook_intake_register websocket_accept websocket_close websocket_connect websocket_mock websocket_receive websocket_route websocket_send websocket_server websocket_server_close weekday_name window with_approval_policy with_autonomy_policy with_command_policy with_dynamic_permissions with_execution_policy with_rate_limit worker_trigger workflow.continue_as_new workflow.pause workflow.publish_query workflow.query workflow.receive workflow.respond_update workflow.resume workflow.signal workflow.status workflow.update workflow_clone workflow_commit workflow_diff workflow_execute workflow_graph workflow_insert_node workflow_inspect workflow_policy_report workflow_replace_node workflow_rewire workflow_set_auto_compact workflow_set_context_policy workflow_set_model_policy workflow_set_output_visibility workflow_typed_output_checkpoint workflow_validate write_file write_file_bytes x25519_agree x25519_keypair yaml_parse yaml_stringify yield_now zip_create zip_extract zstd_decode zstd_encode"
+  built_in: "Err Ok abs acos add_assistant add_message add_system add_tool_result add_user addr_of advance_time agent agent_budget agent_chat_loop agent_chat_route_input agent_chat_wait_for_user_tools agent_config agent_daemon_snapshot agent_daemon_step agent_dispatch_tool_batch agent_dispatch_tool_call agent_inject_feedback agent_llm_turn agent_loop agent_name agent_parse_tool_calls agent_preset agent_session_ancestry agent_session_claim_tool_format agent_session_close agent_session_compact agent_session_current_id agent_session_exists agent_session_fork agent_session_fork_at agent_session_inject agent_session_length agent_session_open agent_session_reset agent_session_seed_from_jsonl agent_session_snapshot agent_session_system_prompt agent_session_tool_format agent_session_trim agent_subscribe agent_trace agent_trace_summary agent_turn agent_typed_output_checkpoint agentic_user append_file arch artifact artifact_apply_intent artifact_command_result artifact_context artifact_derive artifact_diff artifact_diff_review artifact_editor_selection artifact_git_diff artifact_patch_proposal artifact_review_decision artifact_select artifact_test_result artifact_verification_bundle artifact_verification_result artifact_workspace_file artifact_workspace_snapshot asin assemble_context assert assert_eq assert_ne asset_root atan atan2 atomic atomic_add atomic_cas atomic_get atomic_set audit_agent await base32_decode base32_encode base64_decode base64_encode base64url_decode base64url_encode basename blake3 bold brotli_decode brotli_encode bytes_concat bytes_eq bytes_from_base64 bytes_from_hex bytes_from_string bytes_len bytes_slice bytes_to_base64 bytes_to_hex bytes_to_string bytes_to_string_lossy camel_to_kebab camel_to_pascal camel_to_snake cancel cancel_graceful capture_stderr_start capture_stderr_take ceil channel channel_select checkpoint checkpoint_clear checkpoint_delete checkpoint_exists checkpoint_get checkpoint_list chunk circuit_breaker circuit_check circuit_record_failure circuit_record_success circuit_reset clear_persona_hooks clear_tool_hooks close_agent close_channel color command_llm_risk_scan command_policy command_policy_pop command_policy_push command_result_scan command_risk_scan composition_binding_manifest composition_crystallization_trace composition_execute composition_search_examples composition_typescript_declarations compute_content_hash connector_call connector_shared_verify_jwt_inline constant_time_eq contains context_pack_manifest context_pack_manifest_parse context_pack_suggestions continue_as_new conversation cookie_delete cookie_parse cookie_round_trip cookie_serialize cookie_sign cookie_verify copy_file correction_query correction_record corrections.query corrections.record cos csv_parse csv_stringify cwd daemon_resume daemon_snapshot daemon_spawn daemon_stop daemon_trigger date_add date_diff date_format date_from_components date_in_zone date_iso date_now date_now_iso date_parse date_to_zone dedup_by delete_file dim dirname duration_days duration_hours duration_minutes duration_ms duration_seconds duration_to_human duration_to_seconds ed25519_keypair ed25519_sign ed25519_verify egress_policy elapsed enable_tracing ends_with entries env env_or eprint eprintln error_category estimate_tokens eval_metric eval_metrics eval_pack_manifest eval_pack_run eval_suite_manifest eval_suite_run event_log.emit event_log.latest event_log.subscribe event_log_emit exec exec_at execution_root exit exp extname file_exists fixture_user flat_map floor flow_evidence_atom flow_evidence_citation flow_evidence_metadata flow_evidence_transcript flow_invariant_allow flow_invariant_block flow_invariant_confidence flow_invariant_is_blocking flow_invariant_kind flow_invariant_require_approval flow_invariant_warn flow_remediation flow_with_confidence flow_with_evidence flow_with_remediation format friction_clear friction_eval_fixture friction_event friction_events friction_record git.conflicts git.diff git.fetch git.merge_base git.push git.rebase git.repo.discover git.status git.worktree.create git.worktree.remove glob group_by gzip_decode gzip_encode handler_context handoff handoff_context handoff_routes harn.mcp.roots hash_value hex_decode hex_encode hitl_pending hmac_sha1 hmac_sha256 hmac_sha256_base64 home_dir host_call host_capabilities host_has host_mock host_mock_calls host_mock_clear host_mock_pop_scope host_mock_push_scope host_tool_call host_tool_list hostname http_delete http_download http_get http_header http_mock http_mock_calls http_mock_clear http_patch http_post http_put http_request http_response http_response_bytes http_response_json http_response_text http_server http_server_after http_server_before http_server_on_shutdown http_server_readiness http_server_ready http_server_request http_server_route http_server_security_headers http_server_set_ready http_server_shutdown http_server_test http_server_tls_edge http_server_tls_pem http_server_tls_plain http_server_tls_self_signed_dev http_session http_session_close http_session_request http_stream_close http_stream_info http_stream_open http_stream_read invalidate_facts is_cancelled is_err is_infinite is_nan is_ok is_rate_limited is_same is_stderr_tty is_stdin_tty is_stdout_tty is_timeout is_type iter join jq jq_first json_extract json_parse json_pointer json_pointer_delete json_pointer_set json_stringify json_validate jwt_sign jwt_verify kebab_to_camel kebab_to_snake keys len list_agents list_dir list_providers_native llm_available_providers llm_budget llm_budget_remaining llm_call llm_call_safe llm_call_structured llm_call_structured_result llm_call_structured_safe llm_compare_costs llm_completion llm_config llm_cost llm_format_usd llm_healthcheck llm_infer_provider llm_info llm_known_models llm_mock llm_mock_calls llm_mock_clear llm_mock_pop_scope llm_mock_push_scope llm_model_defaults llm_model_info llm_model_tier llm_pick_model llm_pricing llm_provider_catalog llm_providers llm_qc_default_model llm_rate_limit llm_resolve_model llm_resolved_options llm_session_cost llm_stream llm_stream_call llm_usage ln load_run_tree load_skill log log10 log2 log_debug log_error log_info log_json log_set_level log_warn lowercase lowercase_first mailbox_close mailbox_lookup mailbox_metrics mailbox_open mailbox_receive mailbox_send mailbox_try_receive max mcp_call mcp_connect mcp_disconnect mcp_elicit mcp_ensure_active mcp_get_prompt mcp_list_prompts mcp_list_resource_templates mcp_list_resources mcp_list_tools mcp_prompt mcp_read_resource mcp_registry_status mcp_release mcp_report_progress mcp_resource mcp_resource_template mcp_roots mcp_serve mcp_server_card mcp_server_info mcp_tools md5 mean median merge_captain_agent metadata_entries metadata_get metadata_refresh_hashes metadata_resolve metadata_save metadata_set metadata_stale metadata_status metrics_inc microcompact min mkdir mock_stdin mock_time mock_tty monitor_wait_for_native monotonic_ms month_name move_file multipart_field_bytes multipart_field_text multipart_form_data multipart_parse now_ms oncall_captain_agent pair parallel_race parse_junit_xml partition pascal_to_camel pascal_to_snake path_basename path_extension path_is_absolute path_is_relative path_join path_normalize path_parent path_parts path_relative_to path_segments path_stem path_to_native path_to_posix path_with_extension path_with_stem path_workspace_info path_workspace_normalize percentile persona_eval_ladder_manifest persona_eval_ladder_run pg_close pg_connect pg_execute pg_mock_calls pg_mock_pool pg_pool pg_query pg_query_one pg_transaction pid plan_artifact plan_entries platform pow print println progress project_catalog_native project_enrich_native project_fingerprint project_root project_scan_native project_scan_tree_native project_walk_tree_native prompt_mark_rendered prompt_user provider_capabilities provider_capabilities_clear provider_capabilities_install provider_register query_parse query_stringify random random_choice random_int random_shuffle range read_file read_file_bytes read_file_result read_line read_lines read_stdin receive regex_captures regex_match regex_replace regex_replace_all regex_split register_persona_hook register_step_hook register_tool_hook release_captain_agent render render_always_on_catalog render_prompt render_string render_with_provenance repair_agent replace resume_agent review_captain_agent rng_seed round routing_policy run_record run_record_diff run_record_eval run_record_eval_suite run_record_fixture run_record_load run_record_save runtime_context runtime_context_clear runtime_context_get runtime_context_set runtime_context_values runtime_paths sandbox_active_backend sandbox_active_profile sandbox_backend_available scan_directory schema_check schema_expect schema_extend schema_from_json_schema schema_from_openapi_schema schema_is schema_of schema_omit schema_parse schema_partial schema_pick schema_recover schema_to_json_schema schema_to_openapi_schema scripted_user secret_get secret_scan select_artifacts_adaptive self_review send send_input session_cookie session_from_cookies session_sign session_verify set set_add set_color_mode set_contains set_difference set_intersect set_is_disjoint set_is_subset set_is_superset set_remove set_symmetric_difference set_union sha224 sha256 sha384 sha3_256 sha3_512 sha512 sha512_256 shared_cas shared_cell shared_get shared_map shared_map_cas shared_map_delete shared_map_entries shared_map_get shared_map_set shared_map_snapshot shared_metrics shared_scope_id shared_set shared_snapshot shell shell_at sign signed_url simulated_user_post_turn simulated_user_read_tools simulated_user_respond simulated_user_status sin skill_count skill_define skill_describe skill_find skill_list skill_registry skill_remove skill_render skill_select skill_who_signed skills_catalog_entries sleep sleep_ms snake_to_camel snake_to_kebab snake_to_pascal source_dir spawn_agent split sqrt sse_close sse_connect sse_event sse_mock sse_receive sse_server_cancel sse_server_cancelled sse_server_close sse_server_disconnected sse_server_flush sse_server_heartbeat sse_server_mock_disconnect sse_server_mock_receive sse_server_response sse_server_send sse_server_status starts_with stat stddev store_clear store_delete store_get store_list store_save store_set str_pad stream.broadcast stream.collect stream.debounce stream.filter stream.first stream.fold stream.interleave stream.map stream.merge stream.race stream.scan stream.take stream.take_until stream.tap stream.throttle stream.zip sub_agent_request sub_agent_run substring summary_agent supervisor_events supervisor_metrics supervisor_start supervisor_state supervisor_stop sync_gate_acquire sync_metrics sync_mutex_acquire sync_release sync_rwlock_acquire sync_semaphore_acquire tan tar_create tar_extract task_current temp_dir testbench_clock_leaks testbench_fs_diff testbench_is_active throw_error tiktoken_count_tokens tiktoken_tokenizer_info timer_end timer_start timestamp title_case to_float to_int to_list to_string toml_parse toml_stringify tool_bind tool_count tool_def tool_define tool_describe tool_find tool_format_result tool_list tool_parse_call tool_prompt tool_ref tool_registry tool_remove tool_schema tool_select tool_surface_validate tool_synth_invoke tool_synthesis_cache tool_synthesis_clear tool_synthesize trace_end trace_id trace_spans trace_start trace_summary transcript transcript_abandon transcript_add_asset transcript_archive transcript_assets transcript_auto_compact transcript_compact transcript_events transcript_events_by_kind transcript_export transcript_fork transcript_from_messages transcript_id transcript_import transcript_messages transcript_render_full transcript_render_visible transcript_reset transcript_resume transcript_stats transcript_summarize transcript_summary transport_mock_calls transport_mock_clear trigger_fire trigger_inspect_action_graph trigger_inspect_dlq trigger_inspect_lifecycle trigger_list trigger_register trigger_replay trigger_test_harness trim trust.policy_for trust.query trust.record trust.score trust.verify_chain trust_graph_policy_for trust_graph_query trust_graph_record trust_graph_verify_chain trust_query trust_record try_receive type_of unicode_graphemes unicode_normalize unmock_stdin unmock_time unmock_tty unreachable unwrap unwrap_err unwrap_or uppercase uppercase_first url_build url_decode url_encode url_parse user_tools username uuid uuid_nil uuid_parse uuid_v5 uuid_v7 values variance verify_agent verify_signed_url vision_ocr wait_agent waitpoint_cancel waitpoint_complete waitpoint_create waitpoint_wait walk_dir webhook_intake_deregister webhook_intake_feed webhook_intake_list webhook_intake_recent webhook_intake_register websocket_accept websocket_close websocket_connect websocket_mock websocket_receive websocket_route websocket_send websocket_server websocket_server_close weekday_name window with_approval_policy with_autonomy_policy with_command_policy with_dynamic_permissions with_execution_policy with_rate_limit worker_trigger workflow.continue_as_new workflow.pause workflow.publish_query workflow.query workflow.receive workflow.respond_update workflow.resume workflow.signal workflow.status workflow.update workflow_clone workflow_commit workflow_diff workflow_execute workflow_graph workflow_insert_node workflow_inspect workflow_policy_report workflow_replace_node workflow_rewire workflow_set_auto_compact workflow_set_context_policy workflow_set_model_policy workflow_set_output_visibility workflow_typed_output_checkpoint workflow_validate write_file write_file_bytes x25519_agree x25519_keypair yaml_parse yaml_stringify yield_now zip_create zip_extract zstd_decode zstd_encode"
 };
